# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/api/v1/lessons": {
      "get": {
        "operationId": "LessonsController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "coach",
                "student"
              ],
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Filter by lesson date (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instructorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resortId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List lessons with optional filters",
        "tags": [
          "lessons"
        ]
      },
      "post": {
        "operationId": "LessonsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new lesson with seats",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/api/v1/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Lesson ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get lesson details by ID",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/api/v1/lessons/{id}/seats": {
      "get": {
        "operationId": "LessonsController_getSeats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Lesson ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Comma-separated includes (e.g. \"self_eval\")",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get seats for a lesson with status and optional self-evaluation",
        "tags": [
          "lessons"
        ]
      }
    },
...
    "/api/api/v1/invitations/{code}": {
      "get": {
        "operationId": "InvitationsController_verifyCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "8-character invitation code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify invitation code validity",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/api/v1/invitations/claim": {
      "post": {
        "operationId": "InvitationsController_claimSeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Claim a seat using invitation code and student information",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/api/v1/invitations/{code}/identity": {
      "post": {
        "operationId": "InvitationsController_submitIdentityForm",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "8-character invitation code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit or update identity form before claiming seat",
        "tags": [
          "invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "DIY Ski Evaluation API",
    "description": "Ski instructor teaching evaluation and record management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateLessonDto": {
        "type": "object",
        "properties": {}
      },
...
      "ClaimInvitationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}